basic index file checksum verification (#293)
