+++ git-index/Cargo.toml
@@ -30,7 +30,7 @@ internal-testing-to-avoid-being-run-by-cargo-test-all = []
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-git-features = { version = "^0.19.1", path = "../git-features", features = ["rustsha1"] }
+git-features = { version = "^0.19.1", path = "../git-features", features = ["rustsha1", "progress"] }
 git-hash = { version = "^0.9.1", path = "../git-hash" }
 git-bitmap = { version = "^0.0.1", path = "../git-bitmap" }
